+++ ingest/fhir/com.ibm.streamsx.health.fhir/service.properties
@@ -1,9 +1,15 @@
 # base url of the FHIR server
-baseurl=http://fhirtest.uhn.ca/baseDstu3
+# This is the base URL for Hapi Fhir public server
+# baseurl=http://fhirtest.uhn.ca/baseDstu3
+baseurl=
+
 
 # patient Ids to monitor for
-patientids=test-1796238, 191025, 1c6299f7-2d06-4b5d-9efb-a8216a405a92
-
+# These are patient ids available from Hapi Fhir public server available
+# at the time of development.  They may no longer be available.
+# Check the server for a list of patients you can monitor
+# patientids=test-1796238, 191025, 1c6299f7-2d06-4b5d-9efb-a8216a405a92
+patientids=
 # period between each fetch in seconds, default is 10
 period=
 
@@ -17,8 +23,8 @@ debug=
 streamscontext=DISTRIBUTED
 
 # Add VM arguments to JVM, default is empty
-vmargs=
-
 # The following is the example vm args to enable debugging for the service
 # This argument enable the JVM debug engine, allowing debug client to attach to the process remotely
-# vmargs=-agentlib:jdwp=transport=dt_socket,suspend=y,server=y,address=127.0.0.1:7777
+#vmargs=-agentlib:jdwp=transport=dt_socket,suspend=y,server=y,address=127.0.0.1:7777
+vmargs=
+
