+++ ingest/fhir/com.ibm.streamsx.health.fhir/src/com/ibm/streamsx/health/fhir/model/ParseError.java
@@ -1,3 +1,8 @@
+//*******************************************************************************
+//* Copyright (C) 2017 International Business Machines Corporation
+//* All Rights Reserved
+//*******************************************************************************
+
 package com.ibm.streamsx.health.fhir.model;
 
 import java.io.Serializable;
